Feature: Added Synonyms service for getting synonyms of words. Released new candidate api. Minor changes in datatypes of some tables. Added scoring engine for candidates. Event triggering for score added too. Also, fixed the view all documents

+++ src/main/java/com/ashar/job/recruitment/management/Dto/JobOpeningDto.java
@@ -27,7 +27,7 @@ public class JobOpeningDto {
     private Date startDate;
     @NotNull(message = "End date cannot be empty.")
     private Date endDate;
-    private Integer preferredExperience;
+    private Double preferredExperience;
     private List<String> requiredSkills;
     private List<String> additionalSkills;
